server/requirements.txt
pyngrok==5.1.0
torchvision==0.13.0
Werkzeug==2.0.3
Flask==2.0.3
flask-restx==0.5.1
Pillow==9.0.1
torch==1.12.0